+++ ktor-io/nonJvm/src/io/ktor/utils/io/core/Closeable.nonJvm.kt
