+++ ktor-io/posix/src/io/ktor/utils/io/IODispatcher.posix.kt
@@ -1,11 +1,9 @@
 /*
- * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2026 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package io.ktor.utils.io
 
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
+import kotlinx.coroutines.*
 
 public actual fun ioDispatcher(): CoroutineDispatcher = Dispatchers.IO
