--- ktor-io/posix/src/io/ktor/utils/io/core/Closeable.posix.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright 2014-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package io.ktor.utils.io.core
-
-public actual interface Closeable : AutoCloseable {
-    override fun close()
-}
