+++ ktor-io/wasmWasi/src/io/ktor/utils/io/core/ByteOrder.wasmWasi.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright 2014-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.utils.io.core
+
+public actual enum class ByteOrder {
+    BIG_ENDIAN,
+    LITTLE_ENDIAN;
+
+    public actual companion object {
+        public actual fun nativeOrder(): ByteOrder = LITTLE_ENDIAN
+    }
+}
