--- ktor-io/web/src/io/ktor/utils/io/ByteChannel.web.kt
@@ -1,8 +1,0 @@
-/*
- * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package io.ktor.utils.io
-
-internal actual val DEVELOPMENT_MODE: Boolean
-    get() = false
